Reduce number of warnings raised by bugprone-* clang-tidy checks in tracking blocks

+++ src/algorithms/signal_source/gnuradio_blocks/unpack_2bit_samples.cc
@@ -72,8 +72,8 @@ void swapEndianness(int8_t const *in, std::vector<int8_t> &out, size_t item_size
 
     for (i = 0; i < ninput_items; ++i)
         {
-            k = j + skip;
-            l = j;
+            k = static_cast<int>(j + skip);
+            l = static_cast<int>(j);
             while (k >= l)
                 {
                     out[j++] = in[k--];
